requirements.txt
coreapi==2.0.8
Django==1.10.2
django-debug-toolbar==1.6
djangorestframework==3.5.1
djangorestframework-jwt==1.8.0
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.6.0
itypes==1.1.0
mysqlclient==1.3.9
PyJWT==1.4.2
requests==2.11.1
sqlparse==0.2.2
uritemplate==3.0.0
